Moved limitDatePage into LocationOfGoodsNavigator
